# Fix JSON rendering optimization